+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Use Java 21 as the runtime and language level
 * Strimzi Drain Cleaner updated to 1.5.0 (included in the Strimzi installation files)
+* Support for Linux user namespaces in Strimzi Pods through the `hostUsers` Pod option
 
 ### Major changes, deprecations, and removals